Separate getTypeKeys and isObject, further work on validateMapping
