+++ src-runtime/index.mjs
@@ -3,13 +3,16 @@ export * from './createDiv.mjs';
 export * from './customTypes.mjs';
 export * from './customValidations.mjs';
 export * from './div.mjs';
+export * from './getTypeKeys.js';
 export * from './inspectType.mjs';
+export * from './isObject.js';
 export * from './makeJSDoc.mjs';
 export * from './options.mjs';
 export * from './partition.js';
 export * from './registerClass.mjs';
 export * from './registerTypedef.mjs';
 export * from './registerVariable.js';
+export * from './replaceType.js';
 export * from './validateArray.mjs';
 export * from './validateDivision.mjs';
 export * from './validateIntersection.js';
