+++ src-runtime/validateKeyof.mjs
@@ -1,30 +1,4 @@
-import {variables} from "./registerVariable.js";
-import {typedefs } from "./registerTypedef.mjs";
-/**
- * @example
- * getTypeKeys({type: 'typeof', argument: 'DataTypeMap'}, console.warn);
- * @param {*} expect - The type.
- * @param {console["warn"]} warn - Function to warn with.
- * @returns {string[]|undefined} Keys or undefined.
- */
-function getTypeKeys(expect, warn) {
-  if (expect?.type === 'typeof') {
-    const varName = expect.argument;
-    const variable = variables[varName];
-    if (variable === undefined) {
-      warn(`Can't find variable named '${varName}'.`);
-      return;
-    }
-    return Object.keys(variable);
-  } else if (typedefs[expect]) {
-    const typedef = typedefs[expect];
-    if (typedef.type === 'union') {
-      return typedef.members;
-    }
-    warn("getTypeKeys: Unhandled typedef", {expect, typedef});
-  }
-  warn(`Couldn't get keys for type`, expect);
-}
+import {getTypeKeys} from "./getTypeKeys.js";
 /**
  * @param {*} value - The actual value that we need to validate.
  * @param {*} expect - The supposed type information of said value.
